gralloc: Fix mismatched metadata handling
Auto components have MD_PP_PARAM_INTERLACED and
MD_LINEAR_FORMAT metadata types instead of
PP_PARAM_INTERLACED and LINEAR_FORMAT.
So, update this handling correctly in composer and mapper.

Also fix 1 format mismatch while getting USAGE in HWC.

(cherry picked from commit a925caca6f9734d4029ac056c55c929cc81d9e6a)

+++ composer-aidl/hwc_display.cpp
@@ -708,7 +708,7 @@ void HWCDisplay::BuildLayerStack() {
       }
       // TZ Protected Buffer - L1
       // Gralloc Usage Protected Buffer - L3 - which needs to be treated as Secure & avoid fallback
-      int32_t handle_flags;
+      uint64_t handle_flags;
       buffer_allocator_->GetMetadataValue((void *)handle, SnapMetadataType::USAGE, &handle_flags,
                                           sizeof(handle_flags));
      if (handle_flags & qtigralloc::PRIV_FLAGS_SECURE_BUFFER) {
